Restructure tests to mirror source
