+++ robots.txt
@@ -1,6 +1,6 @@
 User-agent: *
 Allow: /
-Disallow: /__tests__/
+Disallow: /tests/
 Disallow: /node_modules/
 
 Sitemap: https://mihaicristiancondrea.github.io/profile/sitemap.xml
