+++ tests/features/songs/presentation/SongsPage.test.ts
@@ -1,7 +1,7 @@
 const PLACEHOLDER_SRC = 'images/placeholder.png';
-const SONG_DATA_SCRIPT = '../src/features/songs/data/SongApiClient';
-const SONG_DOMAIN_SCRIPT = '../src/features/songs/domain/SongMapper';
-const SONG_PRESENTATION_SCRIPT = '../src/features/songs/presentation/SongsPage';
+const SONG_DATA_SCRIPT = '../../../../src/features/songs/data/SongApiClient';
+const SONG_DOMAIN_SCRIPT = '../../../../src/features/songs/domain/SongMapper';
+const SONG_PRESENTATION_SCRIPT = '../../../../src/features/songs/presentation/SongsPage';
 
 function loadSongsModule() {
     require(SONG_DATA_SCRIPT);
